List tests moved out of DataTest
